+++ InstantiateComponents/Assets/InstantiateComponents/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## [1.0.1] - 2024-12-29 - In Development
 ### Modified
-- Modified the name of the parent object of the instance. 
+- Modified the name of the parent object of the instance.
+- Modified the inspector display.
 
 ## [1.0.0] - 2024-12-26
 ### Added